pyproject.toml
[project]
name = "am"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.14",
    "pyyaml>=6.0.2",
    "uvicorn>=0.35.0",
]
